Fix CR prompt with --auth-user-pass & support reading response.
Fixes a bug where there is no prompt for a dynamic (or static) challenge
response when combining "--auth-retry interact" (or --static-challenge)
with --auth-user-pass and the password is read from the file. It also
extends the reading from the file to allow a response to be provided on
the line after the password (which is nice for "push", "phone", etc.).

NOTE: indentation left too deep on one big section so that changes are
clearly seen.