fix cppcheck findings:
(both master and release/2.3)

[src/openvpn/error.c:736]: (error) Possible null pointer dereference: null
[src/openvpn/error.c:736]: (error) Null pointer dereference

function "crash" is not used anywhere. so, it's better to remove it.

(only master)

[src/openvpnserv/interactive.c:601]: (error) Memory pointed to by 'addr_row' is freed twice.
[src/openvpnserv/interactive.c:700]: (error) Memory pointed to by 'fwd_row' is freed twice.
[src/openvpnserv/interactive.c:1329]: (error) Common realloc mistake: 'handles' nulled but not freed upon failure

RemoveListItem free addr_row and fwd_row, so there's no need to free them once again after "goto". It is dirty fix, I'd say interactive.c is overcomplicated here (lots of "if"s and "goto"s), so it is better to rewrite it in some straightforward way.

as for realloc, it is common not to handle failures, it is not very dangerous, it is rather some sugar to stop cppcheck from noising.