Add an option to filter options received from server
Usage: --pull-filter accept|reject "option string"

Permit a client to selectively accept or reject options pushed
by the server. May be used multiple times. The filters are
applied in the order specified to each pushed option received. The
filtering stops as soon as a match is found. Full regex support is
not available but partial matching is used to allow some flexibility.
Thus

  pull-filter accept "ifconfig 10.9.0."
  pull-filter reject "ifconfig "
  pull-filter accept "route 10."
  pull-filter reject "route "

will reject assigned ip unless its in the "10.9.0.0/24" range and
all pushed routes except those starting with "10.". Note the space
at the end of "route " to not reject "route-gateway", for example.

There is an implicit 'pull-filter accept ""' at the end so that all
options not rejected by any filter are accepted.

Acknowledges shameless imitation of --push-remove.
Inspired by Trac #682.

Signed-off-by: Selva Nair <selva.nair@gmail.com>