Update function open_tun: don't stuck on name tun and tap
When using dynamic unit number for new tun device, trying to get the first free number from the kernel driver at first (by setting ppa = -1); then search for it if that failed. This changed is also to make sure the var 'ptr' is assigned to 'dev', because the later code "if ( *ptr == '\0' )" will always dereferencing the pointer.