Update .mailmap to unify and clean up odd names and e-mail addresses
To make the best use of this, have a look at the --use-mailmap option.
In particular git-log and git-shortlog makes use of it.  Also search
for mailmap in the git-log man page, for more ways to use this
remapping (format strings to --pretty).

Otherwise, to make use of the mailmap remapping by default do:

     $ git config [--global] log.mailmap true